added utils class for tasks with tests
